introduce an async version of stratum server. (#2468)
It seems that current approach stops workring after amount of parallel
connection exceeds ~500. This PR introduces an 'async' feature which
enable asyncronius, tokio based implementation of stratum server.

* fix bug with passing current_difficulty in Handler

I passed in Handler current_difficulty as an u64, so it was copied
and all later changes were ignored. In this commit I pass referece on
RwLock with current_difficulty.

* fix crash in build_block

* improve  error handling

* print error message on unknown error

* replace original server by tokio based one

* fixes after review

* reduce sleep time in mail loop to 5ms