Proof of Work Hard Fork
A proof of work change to remove C32 and C33 from activating along with some other changes to strengthen the network.

Because of the consistent and relentless malicious miner attacks, that have only grown in severity and sophistication, we feel these changes are important changes for the viability and strength of the MWC network. There has been significant discussion and consensus among the largest MWC holders, miners and developers regarding these changes.

All major stakeholders are in agreement with these changes to strengthen the network and they have been tested extensively both in individual tests and on testnet. The major stakeholders thought it would be appropriate that I, as the largest single MWC miner who has helped defend against these malicious miner attacks, submit the official pull request.